fix(caching): fix caching
accidentally, instances were cached by key ID. what this meant was that if there was multiple
instances for a particular class, only the latest instance would be always returned, because the
cache was searching only for classID -> instance.

This was fixes with another layer of caching,
so now we first search for class-specific cache and then search that cache for instance. In other
words classID -> {instanceID -> instance }.
Since we also use another layer to reference the
instances only weakly, the full cache structure is {classID -> {instanceID -> instanceProxy}} and
another cache which is {instanceProxy -> instance}