fix(logging): fix package name detection by debug
currently the working dirctory for detecting the name of the package that should be used in debug package was set to current working directory. This worked fine when working in the package itself, but when the package was imported, the cwd pointed elsewhere, and so all debug statements of the first package were included in the cwd's package's debug statements too.

This commit fixes that by specifying to search for the package name in the directory where the file that creates the debug instance is located. So even if the package is imported, the imported package's directory is searched, not the importee's directory.